requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
pytest-recording==0.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
vcrpy==4.2.1
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0